+++ x-pack/plugins/security_solution/public/resolver/view/edge_line.tsx
@@ -9,7 +9,8 @@ import styled from 'styled-components';
 import { FormattedMessage } from '@kbn/i18n/react';
 import { applyMatrix3, distance, angle } from '../models/vector2';
 import { Vector2, Matrix3, EdgeLineMetadata } from '../types';
-import { useResolverTheme, calculateResolverFontSize } from './assets';
+import { fontSize } from './font_size';
+import { useColors } from './use_colors';
 
 interface StyledEdgeLine {
   readonly resolverEdgeColor: string;
@@ -19,7 +20,7 @@ interface StyledEdgeLine {
 const StyledEdgeLine = styled.div<StyledEdgeLine>`
   position: absolute;
   height: ${(props) => {
-    return `${calculateResolverFontSize(props.magFactorX, 12, 8.5)}px`;
+    return `${fontSize(props.magFactorX, 12, 8.5)}px`;
   }};
   background-color: ${(props) => props.resolverEdgeColor};
 `;
@@ -87,8 +88,8 @@ const EdgeLineComponent = React.memo(
      */
     const screenStart = applyMatrix3(startPosition, projectionMatrix);
     const screenEnd = applyMatrix3(endPosition, projectionMatrix);
-    const [magFactorX] = projectionMatrix;
-    const { colorMap } = useResolverTheme();
+    const [xScale] = projectionMatrix;
+    const colorMap = useColors();
     const elapsedTime = edgeLineMetadata?.elapsedTime;
 
     /**
@@ -96,7 +97,7 @@ const EdgeLineComponent = React.memo(
      * should be the same as the distance between the start and end points.
      */
     const length = distance(screenStart, screenEnd);
-    const scaledTypeSize = calculateResolverFontSize(magFactorX, 10, 7.5);
+    const scaledTypeSize = fontSize(xScale, 10, 7.5);
 
     const style = {
       left: `${screenStart[0]}px`,
@@ -120,8 +121,8 @@ const EdgeLineComponent = React.memo(
     /**
      * Calculates a fractional offset from 0 -> 5% as magFactorX decreases from 1 to a min of .5
      */
-    if (magFactorX < 1) {
-      const fractionalOffset = (1 / magFactorX) * ((1 - magFactorX) * 10);
+    if (xScale < 1) {
+      const fractionalOffset = (1 / xScale) * ((1 - xScale) * 10);
       elapsedTimeLeftPosPct += fractionalOffset;
     }
 
@@ -130,7 +131,7 @@ const EdgeLineComponent = React.memo(
         className={className}
         style={style}
         resolverEdgeColor={colorMap.resolverEdge}
-        magFactorX={magFactorX}
+        magFactorX={xScale}
         data-test-subj="resolver:graph:edgeline"
       >
         {elapsedTime && (
