+++ x-pack/plugins/security_solution/public/resolver/view/process_event_dot.tsx
@@ -12,12 +12,15 @@ import { FormattedMessage } from '@kbn/i18n/react';
 import { NodeSubMenu } from './submenu';
 import { applyMatrix3 } from '../models/vector2';
 import { Vector2, Matrix3, ResolverState } from '../types';
-import { SymbolIds, useResolverTheme, calculateResolverFontSize } from './assets';
 import { ResolverEvent, SafeResolverEvent } from '../../../common/endpoint/types';
 import { useResolverDispatch } from './use_resolver_dispatch';
 import * as eventModel from '../../../common/endpoint/models/event';
 import * as selectors from '../store/selectors';
 import { useNavigateOrReplace } from './use_navigate_or_replace';
+import { fontSize } from './font_size';
+import { useCubeAssets } from './use_cube_assets';
+import { useSymbolIDs } from './use_symbol_ids';
+import { useColors } from './use_colors';
 
 interface StyledActionsContainer {
   readonly color: string;
@@ -108,6 +111,8 @@ const UnstyledProcessEventDot = React.memo(
     // This should be unique to each instance of Resolver
     const htmlIDPrefix = `resolver:${resolverComponentInstanceID}`;
 
+    const symbolIDs = useSymbolIDs();
+
     /**
      * Convert the position, which is in 'world' coordinates, to screen coordinates.
      */
@@ -191,7 +196,7 @@ const UnstyledProcessEventDot = React.memo(
      *  18.75 : The smallest readable font size at which labels/descriptions can be read. Font size will not scale below this.
      *  12.5 : A 'slope' at which the font size will scale w.r.t. to zoom level otherwise
      */
-    const scaledTypeSize = calculateResolverFontSize(xScale, 18.75, 12.5);
+    const scaledTypeSize = fontSize(xScale, 18.75, 12.5);
 
     const markerBaseSize = 15;
     const markerSize = markerBaseSize;
@@ -212,7 +217,7 @@ const UnstyledProcessEventDot = React.memo(
           })
         | null;
     } = React.createRef();
-    const { colorMap, cubeAssetsForNode } = useResolverTheme();
+    const colorMap = useColors();
     const {
       backingFill,
       cubeSymbol,
@@ -220,7 +225,7 @@ const UnstyledProcessEventDot = React.memo(
       isLabelFilled,
       labelButtonFill,
       strokeColor,
-    } = cubeAssetsForNode(
+    } = useCubeAssets(
       isProcessTerminated,
       /**
        * There is no definition for 'trigger process' yet. return false.
@@ -323,7 +328,7 @@ const UnstyledProcessEventDot = React.memo(
         >
           <StyledOuterGroup>
             <use
-              xlinkHref={`#${SymbolIds.processCubeActiveBacking}`}
+              xlinkHref={`#${symbolIDs.processCubeActiveBacking}`}
               fill={backingFill} // Only visible on hover
               x={-15.35}
               y={-15.35}
@@ -334,7 +339,7 @@ const UnstyledProcessEventDot = React.memo(
             />
             {isOrigin && (
               <use
-                xlinkHref={`#${SymbolIds.processCubeActiveBacking}`}
+                xlinkHref={`#${symbolIDs.processCubeActiveBacking}`}
                 fill="transparent" // Transparent so we don't double up on the default hover
                 x={-15.35}
                 y={-15.35}
